Add a scopeless access token to public Travis config.
It may seem strange to commit an access token to a public repository but this
one has been granted no scopes, so it doesn't have access to *do* anything, and
is simply used to avoid the GH API's unauthed rate limiting when fetching public
assets. This path was chosen on the advice of the GH platform team.