By removing these hard coded restrictions, more correct indentions are possible
Currently these hard coded restrictions prevent some very valid use
cases (will add some screencasts) and limit the possibilities to
implement correct indentation. As far as I’ve tested with it, I did not
find any negative and/or unexpected behaviour with the restrictions
removed, but of course I did not test with all possible languages Atom
supports.

If it turns out there is a language that does have negative effects
from this change, I would argue that the correct place to fix that is
in the specific language package by altering the increaseIndentPattern
and/or (more likely) the decreaseIndentPattern, instead of relying on
these hard coded restrictions.