Measure lines from DOM nodes when they are available
If a DOM node for a certain line is available (i.e. a line has already been
rendered and showed to the user), we now use it to perform all those
measurements useful to position highlights, cursors and whatnot. We still have a
way to guess the position of such elements, which takes advantages of the old
scoped character widths, when a line is not visible on screen. This is just an
approximation, though, and as a result in the next commits we will request a new
animation frame where we can fix the approximated coordinates with real
positions using DOM nodes if they're available.

The design still needs to be molded a bit and I am thinking to have a separate
object which cleanly encapsulates all the measurements concerns. We'll see how
that develops.