Terminate selection drag on text input
On recent versions of OS X Yosemite, there is a delay before we receive
the `mouseup` event when dragging selections with the trackpad. This
means that we’re frequently accidentally inserting text before the
selection process terminates, leading to unexpected selections after the
text insertion. This fixes that behavior and makes the behavior more
straightforward even in the case where the mouse button is remains held
down during text insertion.