Fix config.get for schemas containing objects
The config.get method would not return the default values for the
properties in objects. Now with _.deepExtends, the default values are
the base while any new values overwrite the defaults. This way all
default values appear, including those nested in an object. This fixes
atom/settings-view#386 and fixes atom/settings-view#518.