Add 'New View For File' to file menu
New option in the file menu to create a new tab for the current active
pane item if the item is a TextEditor object. The new item is then set
as the active item. newViewForActiveTextEditor() returns the new
TextEditor or 'undefined' if the active pane item is not a TextEditor.