Clean up max scroll margin code
Previously we were attempting the same computation twice, once
incorrectly when returning the scroll margins as integers and then doing
the same thing in terms of pixels. This just cleans up the original
calculation. It’s *slightly* different because it takes the floor to
continue returning integers, but the behavior is extremely close.