Simplify page-up/page-down
We simply move the cursor up or down by the number of whole lines that
fit on screen, allowing autoscroll to adjust the scroll position. This
causes there to be a margin under the location to which we move the
cursor, but I think it’s better to provide context and keep the
autoscroll experience consistent when using the keyboard.