Stop highlighting at the end of the line
Editors like Sublime and Code do not highlight the entire line when doing a multi-line selection, instead they stop after the last character on that line.

I wanted to implement the same thing on Atom. The way I did it is very hacky and not production ready but it gets the job done. Right now, if you highlight 5 lines, the first and last lines are their own "tile" and all the remaining ones are grouped into a single rectangular tile. Since we want to have each line with a different width, I forced the tile size to 1 and removed the logic that makes it expand all the way.

I'd really like this feature to be in atom, let me know how I can refactor the code to make it shippable. Thanks