:bug: Teach git-repository-provider to recognize .git-files
Git repositories may be contained in a .git directory or a .git file
in the workdir hierarchy, but Atom only recognizes the directory format.
Teach Atom to recognize the filesystem-agnostic Git symbolic link
used by default in many situations including, for example, submodules.

The .git file contains a relative or absolute path to the location of
the real git-dir, preceded by the 8-byte string "gitdir: ".

Here's a console log showing the normal creation of such a symbolic link.

    /tmp $ git init --separate-git-dir foo.git bar
    Initialized empty Git repository in /tmp/foo.git/
    /tmp $ ls
    /tmp $ bar  foo.git
    /tmp $ ls -la bar
    drwxr-xr-x 2 hordp hordp 4096 Sep 18 15:54 .
    drwxr-xr-x 4 hordp hordp 4096 Sep 18 15:54 ..
    -rw-r--r-- 1 hordp hordp   25 Sep 18 15:54 .git
    /tmp $ ls foo.git
    branches  config  description  HEAD  hooks  info  objects  refs
    /tmp $ cat bar/.git
    gitdir: /tmp/foo.git

Fixes #8876