For now, do not create the default Jasmine reporters for an alternative test runner.
In practice, the TerminalReporter seems to have all sorts of funny requirements due
to a transitive dependency on `failure-tree.coffee`, which is something that only exists
in @kevinsawicki's personal fork of `jasmine-node`: https://github.com/kevinsawicki/jasmine-node.git.

To simplify things, this keeps all of the Jasmine-specific stuff in its own file.
This simplifies the contract of a `--test-runner-bootstrap` such that its only requirement
is to produce an object with an `execute()` method.

It would be nice if the `execute()` method could return a `Promise` and have that handled
properly, though that's a little messy because the `spec-bootstrap.coffee` file is currently
written in CoffeeScript (so no async/await), which is in turn called from `atom-application.coffee`.