:bug: Fix invalid shellAppDir on OS X
This resolves the "perfect storm" scenario that resulted when #8829
got merged with `master`. The merge resulted in an invalid shellAppDir
being used, which in turn propagated all the way into script/set-version
and thus the plist update commands failed since it was pointed at a path
which did not exist.

TL;DR: the .app suffix was missing from shellAppDir due to code reordering.

/cc @maxbrunsfeld