Don’t register defaultTarget for keymaps in constructor after all
We’re clearing the views on reset, so this view will end up being
different between test runs. Also, we didn’t previously always create
the workspace view in many cases, and it makes sense to apply that here.

We may want to get the default target for the keymap via a function so
we can defer view creation until one is needed.