Put custom element registration in a separate method
So to avoid registering elements twice. As of today, there's now way of
unregister elements (e.g. because of legit questions like "what happens when
unregistering an element that's being used?").

As a result, I feel the need to move this method somewhere else, as it seems to
have a different lifecycle from AtomEnvironment's one: ideally, I'd put it near
to `document`, which is most likely where it belongs.

When registering `ViewProvider`s, we should probably create elements through
`document.createElement` in order to avoid the need of referring to the element
prototype.