Move async `didStopChangingActiveItem` to `Pane`
Functionality like focusing the newly active text editor lives in the
`Pane` class and needs access to the async `didStop` callback. Move the
functionality into `Pane` and keep the same API for `PaneContainer` by
adding a new subscription similar to its `observeActiveItem`
subscription.

Throttle the async version of `didChangeActiveItem` because debouncing
can lead to a UI that is significantly out-of-date compared with the
state expected by the user. For example, pressing and holding ctrl-tab
(next tab) would keep the event debounced for the duration of the
keypress. Throttling the event ensures most callbacks are still async
but are able to update the UI periodically if events are happening
quickly.