Focus pane items async, improve tab responsiveness
The native `focus` call is the bulk of the time spent in response to
an active item change event. By debouncing the event and always focusing
asynchronously (async is the default for `debounce` callbacks), tab
changing is more responsive because the 'tabs' package typically updates
its UI only after the native `focus` call completes.

For the crazy perf test of pressing and holding next/previous tab
(ctrl-tab on OS X), many unnecessary `focus` calls are avoided because
of debouncing.