Spike out async git repo again
After implementing almost everything else in #9213, I ran into some
sporadic spec failures that I was having trouble reasoning about. This
is an attempt to start over with the class *beginning* with the
functionality that went flaky, to better understand how it is tested
and whether it was being broken by an interaction with something else
in the new implementation, broken intrinsically, or had buggy tests.