Store project directory paths as state on AtomWindow in browser process
Fixes #9574

Previously, we were storing the project directory paths as the
`initialPaths` key in load settings, which were accessed in the browser
process by reading the URL hash. However, this URL hash was not always
available, subjecting us to timing issues when opening multiple files
in the same folder in rapid succession.

We now store the project directory paths directly on AtomWindow
instances on creation, then RPC changes from the render process to the
browser process with a custom code path.

Shout out to :airplane::finnadie:’d @as-cii on this for pairing with me.