Correct config.coffee doc references to `scope`
The scope selector is referred to with the key `scope` on the options
object, not `scopeDescriptor` as the current docs claim. Reference
`scope` correctly.

Mark `options` as optional for `::observe` because the arguments
collection is checked for length, and a length of 2 assumes the second
argument is `callback`. `options` is actually optional.