:checkered_flag: Use PowerShell to get User path
in Windows when installing via Squirrel, replacing reg.exe
that fails to execute if registry editing is disabled in the group policy.

This was causing the user Path to get deleted if registry editing is
disabled.  The workaround that was in place kept the existing path, but
failed to add Atom to it.