Now using decreaseIndentPattern as default for decreaseThisIndentPattern
This way, if decreaseThisIndentPattern is not specified, we fall back
to the previous name for that pattern (decreaseIndentPattern), while
still being able to change the semantics for decreaseIndentPattern.
This is because we never outdent more than one on the
decreaseThisIndentPattern -- which is what the behavior used to be
(and which has some problems when closing more than one scope).