Fix open dialog default behavior on Linux
Open dialog now defaults to the same directory as the open file,
or the user's home directory if there is no file active (such as
when viewing the Settings page). This is the convention for gedit
and other text editors on Linux.

Because of the render/browser process split, this requires async
messages to get the path over. The alternative is passing this
data when calling application:open, but this requires more
extensive changes.