Distinguish initialPaths from pathsToOpen
The initialPaths that are provided when we restore application state
should not be conflated with pathsToOpen. We want to deserialize the
window’s state based on this set of initialPaths, but we don’t
want to open empty buffers for any of the initialPaths that no longer
exist.

Credit to @jrkoenig for this idea. This is a slightly modified version
of the change he made in #10628.