Learn configuration of FS paths for which to skip VCS integration.
This patch allows users to configure paths such as `/mnt/` or `/Volumes/`
which map to network shares or other targets with slow I/O performance.
By skipping VCS integration on these paths, I/O is significantly reduced at
the cost of not having VCS integration and allowing Atom to at least be usable
as an editor when pointed at mounted network shares.

This is a attempt to mitigate some of the effects of #1766.