🏇 Improve perf for clicking below buffer of long line to seek cursor
In cases where the last line of a file is very long and the user clicks below that line to seek to the end of the last line, it can take a very long time due to calculating the position of every single column.

This commit checks the special case of clicking below the buffer and returns the position without going into the whole calculation loop.

Resolves #10769