Set BrowserWindow's background color to white to re-enable sub-pixel AA
In Electron 0.37.3 the default background color has changed from white
to transparent, and as per the sub pixel anti-aliasing specification,
the layer which hosts the glyphs that need to be rendered needs to have
an opaque background.

We have also observed that upon reload everything seems to work smoothly
even without this fix: this probably points in the direction of a
regression in Electron, but in the meantime this fix should suffice to
allow us to proceed.