Avoid using less imports in nested scopes
Apparently, these imports of 'octicon-utf-codes.less' within mixin
definitions cause any subsequent imports of this file to become noops,
because of the way less dedupes imports. The result is that the variables
defined in that file are only available in the nested scope.

This didn't happen in older versions of less because of bugs which
have been fixed since less 2.7.

Signed-off-by: Antonio Scandurra <as-cii@github.com>