Fix release channel detection on Travis
In the previous build scripts, we were considering all the builds
triggered by a pull-request to be part of the dev channel. We adopted
the same heuristic for the new build scripts, but didn't notice that
Travis always sets the `TRAVIS_PULL_REQUEST` variable to "false" (which
causes the `process.env.TRAVIS_PULL_REQUEST` expression to be evaluated
as truthy).

In the past, this wasn't a problem because we were building artifacts
via Janky, but after switching to Travis this makes it generate the
wrong assets on stable and beta.

To fix this, it seems reasonable to remove the conditional that checks
if we are building a pull-request: in the past this could have been
problematic because assets could be uploaded inadvertently to S3 or
GitHub, but this should be safe now that we rely on the release
publisher to perform that task.