Compile native modules on Travis using clang-3.3
All the native modules we ship with Atom are compiled by dynamically
linking the libstdc++ library. Using gcc (and g++) > 4.6 to compile
them, however, requires a newer version of that library, which is not
installed by default on Ubuntu Precise (Travis default image) and thus
causing all the third-party packages built via Travis to fail when using
Atom 1.11-beta.

By using clang-3.3 we can keep backwards compatibility with previous
versions of libstdc++ and still be able to compile C++11 code (which is
mandatory for Node >= 4).