README.md Tranlation to Simplified Chinese
Hello, this is a translation of the README.md to simplified Chinese. I translated this using Google Translate as well as using my understanding towards the language to correct confusing sentences and words. However, it still needs some tweaking. I did not translate “Flight Manual" because there isn't a Chinese version for it that I can point to. 

Thank you.