Prevent method in onWillDestroyPaneItem callback
Added a new key to the object provided to the onWillDestroyPaneItem
callback with the name 'prevent'. The value for this key is a method
that can be used to prevent closing the given pane item.

This is done via a flag in the function calling the callback notifying
that it shouldn't actually destroy the pane item.