Fix destroying a `PanelContainer` containing multiple panels
Previously, when calling `destroy` on a `PanelContainer` containing
multiple panels, Atom would throw a `Cannot read property 'destroy' of
undefined` exception. This was due to iterating over the panels while at
the same time destroying them, which caused the iterated array to be
modified during the loop.

With this commit we slice the array before iterating over it so that
destroying a `PanelContainer` doesn't throw exceptions anymore.