Delete test verifying that `GitRepository.open` works with file paths
With the upgrade of git-utils libgit2 was updated too and the behavior
tested by the deleted spec doesn't seem to be supported anymore.
However, we believe we can delete this test because the only entry point
for creating `GitRepository` instances is `repositoryForDirectorySync`
in `GitRepositoryProvider`, which only allows `Directory` objects to be
supplied as its input arguments.