penguin: Respect DISPLAY env variable in Unix like environments
Atom connects to a socket to determine if it is already running
and if it should launch a new window in the same process.
This socket is named with the user's name but ignores the user's
display. If a user has multiple displays, such as running a
remote display server, Atom will pop up the new window on the display
that it was first started on.

This patch adds the display name to the socket name if it exists
in the env. Because a seperate socket is being opened Atom does
the correct thing and launches on the display it was started on.