Fix folding key bindings
The documentation for Linux and Windows says that you need to use
`Ctrl+Alt+F` key binding to fold arbitrary sections of your code or
text. However, in `linux.cson` and `win32.cson` declared a
`Ctrl-Alt-Shift-[` key binding, which leads to incorrect processing of
keystroke and the appearance of duplicates. `Ctrl-Alt-Shift-[` key
binding processed as `Ctrl-Alt-{` key binding which is already declared
to fold everything.