Minor changes to compile-cache.js for atom-in-orbit.
Currently, atom-in-orbit's build script has a step
where it applies these changes locally using a patch file,
which is gross. By upstreaming what atom-in-orbit needs
exported, we can remove this gross hack.

I tested this with the complementary change to atom-in-orbit
and verified that the new build worked properly.

Released under CC0.