Speed up parsing the snapshot source map
When calling remote functions or emitting deprecation warnings
respectively Electron and Grim create a fake `Error` object to retrieve
the stack trace of the current call site.

When doing this for the first time, if the call site was located inside
a snapshotted file, previously we would parse the source map for the
snapshot and translate the position of each call in the stack trace.
However, since the snapshot source map is quite big, we were observing
major slowdowns when parsing it for the first time.

With this commit we will parse the snapshot source map while generating
the snapshot, which will allow to not pay for it during runtime.

Signed-off-by: Michelle Tilley <binarymuse@github.com>