Don't snapshot the `iconv-lite` module
When deciding which methods to export, `iconv-lite` verifies the
availability of some Node features, which is clearly an unsupported
behavior when executing it via `mksnapshot`.

In this case we were not getting any exception at compile-time, but we
noticed some methods were missing from iconv-lite's API. With this
commit we are excluding this module from the snapshot, thus making all
of its methods available again.