Attempt to fix flaky test re: blinking cursor
As shown in #15122, this test sometimes fails in CI with the following
error:

  TextEditorComponent
    rendering
      it blinks cursors when the editor is focused and the cursors are not moving
        Expected '0' to be '1'.
          at it (C:\projects\atom\spec\text-editor-component-spec.js:414:49)
        Expected '0' to be '1'.
          at it (C:\projects\atom\spec\text-editor-component-spec.js:415:49)

I *think* this might be a case of overspecification in the test's
assertions. Prior to this commit, the test expected the blinking cursor
to *start* in the visible state, and then transition to the invisible
state. When we see the failure above, I suspect that the cursor has
already transitioned from the visible state to the invisible state by
the time the assertion runs.

Since the test aims to verify that the cursor blinks, it seems like we
should focus on the blinking, and not worry about the *initial* state of
the cursor. This commit removes the assertions that verify the initial
state of the cursor, and instead asserts that the cursor toggles between
the visible and the invisible state.