:arrow_down: pathwatcher, atom-keymap and tree-view
This partially reverts c8dd400 as we don't want to take the risk of
cherry-picking unrelated changes on stable in order to fix #15267.

Please, note that decrementing the version of these dependencies has the
cost of introducing some duplication in the `node_modules` folder due to
pathwatcher not being deduped, which slightly increments the size of the
final application.