Always assign a project path outside of bundle for legacy package specs
This prevents package specs that don't have a fixtures directory from
attempting to read files out of a non-existent directory inside the ASAR
bundle, which causes ENOTDIR errors in superstring.

If the spec does not have a parent folder containing a fixtures
directory, we now set the default project path to `os.tmpdir()`.