Ensure style updates take effect after the editor is re-attached
Previously, performing a style change while an editor was detached would
cause the editor component to not observe the change. Thus, if the
editor was later re-attached, it would use potentially incorrect
measurements to position lines, cursors, etc.

With this commit we will now keep track of a global counter of how many
times the styles have changed. Also, each editor component will have its
own local counter of how many style changes it has observed. If on
re-attach the global counter is different from the local counter, it
means that the editor component is using stale measurements and, as
such, needs to clear them out and re-measure.