Elevate installation step for Linux during 'script/build --install'
This change fixes #14848 which reports that running 'script/build --install' on Linux
fails with permissions errors.  This happens because the user is advised to run the
build script as a normal user (without sudo) because the script does not work when
run as sudo.

The solution is to use ChildProcess.execSync to invoke Node.js to call into our
installation code with sudo permissions just at the point where installation is
needed.  Windows and macOS paths do not receive similar elevation since it isn't
needed on those platforms.

This change also fixes #7134.