Fix #14716: Error when deserializing TreeView for project
This change fixes an issue that appears when the user removes the last
project from the workspace and then re-adds it.  At this time an error
is thrown from within ItemRegistry because the PaneContainer that
holds the TreeView in the left-most dock does not clear out its
existing ItemRegistry before deserializing the old TreeView state.

The fix is to create a new ItemRegistry when a PaneContainer is
deserialized so that the previous Pane's items are not retained.