Add a new keybinding for window:reload Fixes #1794
While #1794 is closed, the issue is still relevant.

This commit adds a new (and default) keybinding for "View/Developer/Reload Window" that isn't in conflict with 1Password's keybindings. It's not unreasonable to presume that there's a large percentage of 1Password users among Atom developer community, so it would be beneficial fix this binding clash.

I've left previous binding in, so users unaffected by this problem won't have to adjust.