:bug: Fix documentation for the Config class
The parser that generates the documenation requires the comment for the
documentation to be immediately preceeding the class definition. As part
of the decaffeination process in #16593 a variable declaration was
placed between the class definition and it's documentaiton, causing the
parser to think there was none... and excluding it from the final
result.

Fixes #16955.