Convert arguments to absolute paths
This script launches cmd and has it launch atom with the passed arguments. The problem is that cmd will not (necessarily) start in the pwd. So, relative paths like `.` and `..` have a different meaning. So we should first convert them to absolute paths with cygpath. 

Try it yourself. Go to some directory in bash and run `atom .` Rather than opening your current bash pwd, you'll open $HOME because when cmd opens, that's where it will be. Now apply this change and try the same thing. 

By the way, I only see this problem affecting MSysGit and MSys2. I can't say whether it affects other environments.