Reset goal column on all cursor changes
Previously, pressing the home key (move-to-first-character-of-line)
while on an empty line wouldn't clear the goal column. This is because
it was only cleared on cursor *change* and that didn't result in a
change. With this commit, it's *always* cleared. Operations that want
to preserve the goal column can reset it afterwards.