:bug: Replace gconf2 with libgconf-2-4
The `gconf2` and `gconf-service` dependencies originally were added in 
f431bb6 from the WebUpd8 version of the Atom .deb file. It looks like 
although they seemed to work for the most part, they were likely 
incorrect originally, and certainly are now.

This commit switches these over to the `libgconf-2-4 | libgconf2-4` 
dependency listed in electron-installer-debian and Chrome.