deb packaging: Drop python dependency
Python is only needed for apm --> npm --> node-gyp.
(For building Atom packages that include native C/C++ code.)

The rest of Atom/apm works 100%, even with no Python installed.

With Python 2 soon to be dropped from the Debian/Ubuntu repos,
having a hard dependency on `python` or `python2` is a problem.

None of the other OSes/platforms have an install-time requirement of
having Python on the system, so this is in line with Atom packaging
for the other platforms.