:fire: .python-version
.python-version specifies which version(s) of Python are in the PATH,
and which version runs when you run the python, python2, or python3
commands.

That said, it is overly specific for this repository's needs.
In order to specify any version, you must enter it down to the patch
level, e.g. 3.8.2; Major-minor versions aren't allowed, e.g. 2.7

If users want to add such a file on their own machines, they may...
But it's unneccesary for this repository to ship this file as if there
were specific "correct" versions of Python to use.

Any Python 2.7.x or 3.5.0+ will work at the moment.

There are other checks elsewhere in the project, such as in
script/bootstrap. These should be sufficient to inform users which
Python versions they can use. node-gyp will also tell you.