script: Download correct electron binaries
If users manually run `npm install`, we want to
make sure the correct Electron-vendored chromedriver and mksnapshot
are downloaded. (Requires ELECTRON_CUSTOM_VERSION to be set properly.)

This postinstall script sets that var and gets the right binaries,
even if the ELECTRON_CUSTOM_VERSION env var wasn't manually set.

(The bootstrap script already handles this for bootstrapping,
but not for manually running "npm install" in the scripts dir.)