Filter out macOS Gatekeeper flag during startup
We don't need this flag, and yargs thinks it is many short flags.
So, we filter it out before it causes too much trouble.

("Trouble" meaning TypeErrors, partial startup, no windows opened...)

---

Background info:

When launching Atom on macOS, if Gatekeeper/XProtect intercepts the
app launch, and the user clicks through/if launching continues,
Gatekeeper/XProtect add a flag of the form: `-psn_0_[1234567]`.

(The six or seven digits at the end
are apparently a "Process Serial Number",
which the internets say dates back to the old "Carbon" framework...)

We should filter this out at some point, as it's not desirable input.
Ideally, we should filter it out before it gets to yargs.

(Otherwise yargs will see it as twelve or thirteen short flags:
`-p`, `-s`, `-n`, `-_`, `-0`, `-_` (again), `-1`...
which is a lot messier and harder to sort out than a single string.)

---

Fixes: https://github.com/atom/atom/issues/21013 (probably)
Fixes: https://github.com/atom/atom/issues/21751
Fixes: https://github.com/atom/atom/issues/21820