script/test: Allow skipping arbitrary packages
Set `ATOM_PACKAGES_TO_SKIP` in your environment variables
in order to skip running tests for any package(s).
It should be a comma-separated string of package names.

For example: export ATOM_PACKAGES_TO_SKIP="github,about,dalek"

It removes packages from an array of packages
that would otherwise be tested. So it works with other options,
such as the "packages1" and "packages2" options,
and the `ATOM_PACKAGES_TO_TEST` env var.