Use ?? ?. (changes != to !== once)
Line 67 is technically different
Line 584 is only different if String.prototype.match is overridden. The original match function always either returns an array or null
Line 885 is also technically different (example: null)
Line 1005 is also technically different (example: {getPath: false})
and also Line 1314
and also Line 1309, but only with a Proxy (example: new Proxy({}, {call(){return 2}}) --> error vs undefined), which is so obscure that I changed it anyways.
Line 1498 is different though because there's an implicit Boolean conversion because of the if statement
paths && paths.length > 0 ---> Boolean(paths) && paths.length > 0
paths?.length > 0 ---> paths.length == null ? false : paths.length > 0

Theoretically, if paths was falsy, but paths.length was set to 1, it would be different. No good reason to do that for any falsy value


this.project is either truthy or null, so some changes based on that were made
Note to self: Comment out `if (this.project)` on line 453