Use electron.screen to check if `dimensionsAreOnScreen`
I don't know enough about `applicationDelegate`. So I'm just using `screen` because it's so simple.
https://github.com/electron/electron/blob/v9.4.4/docs/api/structures/display.md

But if there's another way feel free to edit.

Using `workArea` instead of `bounds`
Right now, the whole area must be onscreen. But it probably makes more sense to allow, like, 10% of the area to be offscreen.